feat: add verification and retrieval functions for resource limits
